Basic WFC Made
Very crude version with some hard coded (duplicate) lines.
In this version every tile needs 4 lists with all right connections (nto dynamic)
Tiles cant rotate (so need multiple versions of one type)

+++ Assets/Scripts/GridVisualizer.cs
@@ -3,15 +3,15 @@ using UnityEngine;
 
 public class GridVisualizer : MonoBehaviour
 {
-    [Header("Grid")] [SerializeField] private float columns = 5;
-    [SerializeField] private float rows = 5;
+    [Header("Grid")] [SerializeField] private int columns = 5;
+    [SerializeField] private int rows = 5;
     [SerializeField] private Vector3 cellSize = new Vector3(2, 2, 2);
     [SerializeField] private Vector3 startPosition = Vector3.zero;
     [Header("Gizmos")] [SerializeField] private bool drawGizmos = true;
     [SerializeField] private Color color = Color.yellow;
 
-    public float GetColumns => columns;
-    public float GetRows => rows;
+    public int GetColumns => columns;
+    public int GetRows => rows;
     public Vector3 GetCellSize => cellSize;
     public Vector3 GetStartPosition => startPosition;
